relocate to a shared helper
